+++ src/main/java/com/quickhome/entity/HomeInformation.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -19,8 +20,8 @@ public class HomeInformation {
     private String homeDescribe_zch_hwz_gjc;
     private int maxPerson_zch_hwz_gjc;
     private String homeFacilities_zch_hwz_gjc;
-    private double homeDayRent_zch_hwz_gjc;
-    private double homeDeposit_zch_hwz_gjc;
+    private BigDecimal homeDayRent_zch_hwz_gjc;
+    private BigDecimal homeDeposit_zch_hwz_gjc;
     private String homeState_zch_hwz_gjc;
     private String homeAddress_zch_hwz_gjc;
     private Date homeInDate_zch_hwz_gjc;
@@ -29,7 +30,7 @@ public class HomeInformation {
     private String standby3_zch_hwz_gjc;
     private int visible_zch_hwz_gjc;
 
-    public HomeInformation(long homeId_zch_hwz_gjc, String homeName_zch_hwz_gjc, String homeType_zch_hwz_gjc, float homeArea_zch_hwz_gjc, String homeDescribe_zch_hwz_gjc, int maxPerson_zch_hwz_gjc, String homeFacilities_zch_hwz_gjc, double homeDayRent_zch_hwz_gjc, double homeDeposit_zch_hwz_gjc, String homeState_zch_hwz_gjc, String homeAddress_zch_hwz_gjc, Date homeInDate_zch_hwz_gjc) {
+    public HomeInformation(long homeId_zch_hwz_gjc, String homeName_zch_hwz_gjc, String homeType_zch_hwz_gjc, float homeArea_zch_hwz_gjc, String homeDescribe_zch_hwz_gjc, int maxPerson_zch_hwz_gjc, String homeFacilities_zch_hwz_gjc, BigDecimal homeDayRent_zch_hwz_gjc, BigDecimal homeDeposit_zch_hwz_gjc, String homeState_zch_hwz_gjc, String homeAddress_zch_hwz_gjc, Date homeInDate_zch_hwz_gjc) {
         this.homeId_zch_hwz_gjc = homeId_zch_hwz_gjc;
         this.homeName_zch_hwz_gjc = homeName_zch_hwz_gjc;
         this.homeType_zch_hwz_gjc = homeType_zch_hwz_gjc;
